+++ src/shared/PID.hpp
@@ -4,7 +4,9 @@ class PID{
   public:
     PID(float kP, float  kI, float kD, bool feedForward);
     void set(float setPoint);
-    float run(float value);
+
+    /** this returns an int because motor speeds can't be more precise than an int after scaling */
+    int run(float value);
 
     float kP, kI, kD;
     float  lastError,
